Moved ImplementedArrayRolling into rolling
